// src/types/things-iate.ts
import { TrackedFoodItem } from './tracked-food-item';

export class ThingsIAte {
    forBreakfast: TrackedFoodItem[] = [];
    asAMorningSnack: TrackedFoodItem[] = [];
    forLunch: TrackedFoodItem[] = [];
    asAnAfternoonSnack: TrackedFoodItem[] = [];
    forDinner: TrackedFoodItem[] = [];
    forDessert: TrackedFoodItem[] = [];
}